Refactor the ioctl handler of tty module
