+++ kernel/src/device/tty/hvc.rs
@@ -49,10 +49,6 @@ impl TtyDriver for HvcDriver {
 
     fn notify_input(&self) {}
 
-    fn console(&self) -> Option<&dyn AnyConsoleDevice> {
-        Some(&*self.console)
-    }
-
     fn on_termios_change(&self, _old_termios: &CTermios, _new_termios: &CTermios) {}
 }
