+++ kernel/src/process/process/terminal.rs
@@ -36,7 +36,16 @@ mod ioctl_defs {
 }
 
 impl dyn Terminal {
-    pub fn job_ioctl(self: Arc<Self>, raw_ioctl: RawIoctl, via_master: bool) -> Result<()> {
+    /// Handles job-control ioctls.
+    ///
+    /// The return value depends on whether the ioctl is recognized:
+    /// - If the terminal recognizes and handles the ioctl, it should return
+    ///   `Ok(true)`.
+    /// - If an error occurs while processing the ioctl, it should return
+    ///   `Err(_)`.
+    /// - If the terminal does not recognize the ioctl command, it should return
+    ///   `Ok(false)` to indicate that the ioctl command is not supported.
+    pub fn job_ioctl(self: Arc<Self>, raw_ioctl: RawIoctl, via_master: bool) -> Result<bool> {
         use ioctl_defs::*;
 
         dispatch_ioctl!(match raw_ioctl {
@@ -54,7 +63,7 @@ impl dyn Terminal {
                     self.is_control_and(&current!(), |_, _| Ok(operate()))?
                 };
 
-                cmd.write(&pgid)
+                cmd.write(&pgid)?;
             }
             cmd @ SetForegroundPgid => {
                 let pgid = cmd.read()?;
@@ -62,7 +71,7 @@ impl dyn Terminal {
                     return_errno_with_message!(Errno::EINVAL, "negative PGIDs are not valid");
                 }
 
-                self.set_foreground(pgid, &current!())
+                self.set_foreground(pgid, &current!())?;
             }
 
             // Commands about sessions
@@ -71,7 +80,7 @@ impl dyn Terminal {
                     warn!("stealing TTY from another session is not supported");
                 }
 
-                self.set_control(&current!())
+                self.set_control(&current!())?;
             }
             _cmd @ SetControlNoTty => {
                 if via_master {
@@ -81,7 +90,7 @@ impl dyn Terminal {
                     );
                 }
 
-                self.unset_control(&current!())
+                self.unset_control(&current!())?;
             }
             cmd @ GetControlSid => {
                 let sid = if via_master {
@@ -98,14 +107,16 @@ impl dyn Terminal {
                     self.is_control_and(&current!(), |session, _| Ok(session.sid()))?
                 };
 
-                cmd.write(&sid)
+                cmd.write(&sid)?;
             }
 
             // Commands that are invalid or not supported
             _ => {
-                return_errno_with_message!(Errno::ENOTTY, "the ioctl command is unknown")
+                return Ok(false);
             }
-        })
+        });
+
+        Ok(true)
     }
 
     /// Sets the terminal to be the controlling terminal of the process.
